evals: separate actor from participant
